+++ frontend_native/src/renderer.rs
@@ -195,12 +195,12 @@ impl Renderer
         return false;
     }
 
-    pub fn queue_sound(&mut self, queue: &Vec<i16>) {
+    pub fn queue_sound(&mut self, queue: &Vec<f32>) {
         if SOUND_ENABLED && !self.speed_up {
             //println!("queue-size: {}", self.sound_player.device.size());
             if self.sound_player.device.size() == 0 {
                 println!("Audio gap!");
-                self.sound_player.device.queue(&vec![0; 4096]);
+                self.sound_player.device.queue(&vec![0 as f32; 4096]);
             }
             self.sound_player.device.queue(queue);
         }
